+++ lib/rarbg/api.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'faraday'
-require 'faraday_middleware'
+require 'json'
 
 # Main namespace for RARBG.
 module RARBG
@@ -55,8 +55,6 @@ module RARBG
     #   rarbg = RARBG::API.new
     def initialize
       @connection = Faraday.new(url: API_ENDPOINT) do |conn|
-        conn.request  :json
-        conn.response :json, content_type: /\bjson$/
         conn.response :logger if $VERBOSE
         conn.adapter  Faraday.default_adapter
 
@@ -104,10 +102,7 @@ module RARBG
     def list(params = {})
       raise ArgumentError, 'Expected params hash' unless params.is_a?(Hash)
 
-      params.update(
-        mode:  'list',
-        token: token?
-      )
+      params.update(mode: 'list', token: token?)
       call(params)
     end
 
@@ -153,10 +148,7 @@ module RARBG
     def search(params = {})
       raise ArgumentError, 'Expected params hash' unless params.is_a?(Hash)
 
-      params.update(
-        mode:  'search',
-        token: token?
-      )
+      params.update(mode: 'search', token: token?)
       call(params)
     end
 
@@ -202,6 +194,7 @@ module RARBG
       normalize.each_pair do |key, proc|
         params[key] = proc.call(params[key]) if params.key?(key)
       end
+
       params
     end
 
@@ -220,6 +213,7 @@ module RARBG
       SEARCH_KEYS.each do |k|
         params["search_#{k}"] = params.delete(k) if params.key?(k)
       end
+
       params
     end
 
@@ -235,20 +229,22 @@ module RARBG
     # Return or renew auth token.
     def token?
       if token.nil? || time >= (token_time.to_f + TOKEN_EXPIRATION)
-        response = request(get_token: 'get_token')
-        @token = response.fetch('token')
+        response    = request(get_token: 'get_token')
+        @token      = response.fetch('token')
         @token_time = time
       end
+
       token
     end
 
     # Perform API request.
     def request(params)
       rate_limit!(RATE_LIMIT)
-      response = connection.get(nil, params)
+
+      response      = connection.get(nil, params)
       @last_request = time
 
-      return response.body if response.success?
+      return JSON.parse(response.body) if response.success?
 
       raise APIError, "#{response.reason_phrase} (#{response.status})"
     end